@@ -1,14 +1,15 @@
 {
   "name": "kiddy",
-  "description": "Promise wrapper for node mysql connection pool handling",
-  "version": "0.6.0",
-  "main": "index.js",
+  "description": "Promise wrapper for managing a mysql connection pool",
+  "version": "1.0.0-alpha",
+  "exports": "./lib/kitty.js",
+  "type": "module",
   "engines": {
-    "node": ">=6.0.0"
+    "node": ">=12.0.0"
   },
   "author": "Kalen Hall",
   "license": "MIT",
-  "keywords": "promise, mysql, connection-pool",
+  "keywords": "promise, mysql, connection-pool, pooling",
   "homepage": "https://github.com/KalenAnson/kiddy#readme",
   "dependencies": {
     "config": "*",
@@ -31,6 +32,6 @@
     "url": "git+https://github.com/KalenAnson/kiddy.git"
   },
   "devDependencies": {
-    "ava": "^0.25.0"
+    "ava": "^3.15.0"
   }
 }
