@@ -2,28 +2,27 @@
 |*| Kiddy MySQL Connection Pool Handler
 |*| Requires
 \*/
-var Q = require('q');
-var config = require('config');
-var mysql = require('mysql');
-var exports = module.exports;
-var configCache = undefined;
-var pool = undefined;
+import Q from 'q';
+import config from 'config';
+import mysql from 'mysql';
+let configCache = undefined;
+let pool = undefined;
 /*\
 |*| Manually set the configuration options.
 |*| Wrapped in a promise to make this thenable.
 \*/
-exports.setConfig = function (options) {
+export const setConfig = function (options) {
 	return Q.Promise(function (resolve, reject, notify) {
 		/*\
 		|*| Check for manditory stuff
 		\*/
 		if (typeof options === 'undefined') {
-			return reject('Kiddy: missing options');
+			return reject(new Error('Kiddy: missing options') );
 		}
 		if (typeof options.user === 'undefined') {
-			return reject('Kiddy: invalid options, the user option is mandatory');
+			return reject(new Error('Kiddy: invalid options, the user option is mandatory') );
 		}
-		var creds = {
+		let creds = {
 			host: options.host || "localhost",
 			port: options.port || 3306,
 			user: options.user,
@@ -38,7 +37,7 @@ exports.setConfig = function (options) {
 |*| Get a connection handle from a mysql connection pool. If the pool does not
 |*| exist, it will attempt to create the pool.
 \*/
-exports.touch = function () {
+export const touch = function () {
 	return Q.Promise(function (resolve, reject, notify) {
 		_checkConfig()
 		.then(function (creds) {
@@ -55,23 +54,50 @@ exports.touch = function () {
 /*\
 |*| Release a connection handle back to the pool for reuse.
 \*/
-exports.release = function (conn) {
+export const release = function (conn) {
 	return Q.Promise(function (resolve, reject, notify) {
 		if (typeof conn === 'undefined') {
-			return reject('Kiddy: invalid connection handle');
+			return reject(new Error('Kiddy: invalid connection handle') );
 		}
 		if (typeof conn.release === 'undefined') {
-			return reject('Kiddy: unknown connection handle');
+			return reject(new Error('Kiddy: unknown connection handle') );
 		}
 		conn.release();
 		return resolve('Connection released');
 	});
 };
 /*\
+|*| Drain the connection pool.
+\*/
+export const drain = function (opt) {
+	return Q.Promise(function (resolve, reject, notify) {
+		if (typeof opt !== 'undefined' && opt !== false) {
+			_clearConfig();
+		}
+		if (typeof pool === 'undefined') {
+			return resolve('Drained')
+		};
+		if (typeof pool.end === 'function') {
+			Q.ninvoke(pool, 'end')
+			.then(function (status) {
+				pool = undefined;
+				return resolve('Drained')
+			})
+			.fail(function (err) {
+				pool = undefined;
+				return reject(err);
+			});
+		} else {
+			pool = undefined;
+			return reject(new Error('Kiddy: connection pool is invalid') );
+		}
+	});
+};
+/*\
 |*| Get the current mysql connection information. Wrapped in a promise to keep
 |*| the `thenable` program flow used in other database function calls.
 \*/
-_checkConfig = function () {
+const _checkConfig = function () {
 	return Q.Promise(function (resolve, reject, notify) {
 		/*\
 		|*| Resolve the cache or load the config from disk
@@ -80,7 +106,7 @@ _checkConfig = function () {
 			/*\
 			|*| Test for valid configuration
 			\*/
-			var key = "";
+			let key = "";
 			if (config.has('rdb') ) {
 				key = config.rdb;
 			} else if (config.has('database') ) {
@@ -88,12 +114,12 @@ _checkConfig = function () {
 			} else if (config.has('mysql') ) {
 				key = config.mysql;
 			} else {
-				return reject('Kiddy: unable to locate databse configuration in config file');
+				return reject(new Error('Kiddy: unable to locate databse configuration in config file') );
 			}
 			if (typeof key === 'undefined') {
-				return reject('Kiddy: missing database configuration');
+				return reject(new Error('Kiddy: missing database configuration') );
 			} else {
-				var creds = {
+				let creds = {
 					host: key.host || "localhost",
 					port: key.port || 3306,
 					user: key.user,
@@ -111,7 +137,7 @@ _checkConfig = function () {
 /*\
 |*| Create or verify the database pool and retrieve a connection handle.
 \*/
-_touchPool = function (info) {
+const _touchPool = function (info) {
 	return Q.Promise(function (resolve, reject, notify) {
 		if (typeof pool === 'undefined') {
 			try {
@@ -140,7 +166,7 @@ _touchPool = function (info) {
 			});
 		} else {
 			pool = undefined;
-			return reject('Kiddy: connection pool was drained or is invalid');
+			return reject(new Error('Kiddy: connection pool was drained or is invalid') );
 		}
 	});
 };
@@ -148,7 +174,7 @@ _touchPool = function (info) {
 |*| Return the vanilla mysql connection wrapped with the additional method:
 |*| conn.pq - promise wrapped call to conn.query
 \*/
-_wrapConnection = function (conn) {
+const _wrapConnection = function (conn) {
 	return Q.Promise(function (resolve, reject, notify) {
 		if (typeof conn.query === 'function') {
 			conn.pq = function (options, values) {
@@ -156,7 +182,20 @@ _wrapConnection = function (conn) {
 			}
 			return resolve(conn);
 		} else {
-			return reject('Kiddy: invalid msyql connection passed to _wrapConnection');
+			return reject(new Error('Kiddy: invalid msyql connection passed to _wrapConnection') );
 		}
 	});
-}
+};
+const _clearConfig = function () {
+	if (typeof configCache !== 'undefined') {
+		/*\
+		|*| JS does not really have a secure way to remove strings
+		\*/
+		delete configCache.host;
+		delete configCache.port;
+		delete configCache.user;
+		delete configCache.password;
+		delete configCache.database;
+		configCache = undefined;
+	}
+};
